change file for writearticle
edit창과 writearticle이 비슷하기 때문에 같은 코드를 사용하기 위해 writearticle --> component 파일로 옮김
