+++ src/components/index.ts
@@ -7,3 +7,4 @@ export { default as InfoContainer } from './InfoContainer';
 export { default as Profile } from './ProfileBox';
 export { default as StringInput } from './StringInput';
 export { default as Thumbnail } from './Thumbnail';
+export { default as WriteArticle } from './WriteArticle';
